requirements.txt
flask==3.0.3
flask_sqlalchemy==3.1.1
flask_login==0.6.3
flask_mail==0.9.1
python-dotenv==1.0.1
